refactor(network): replace handwritten diagrams with diagrams in asciiflow
