+++ elfo-network/src/socket/capabilities/mod.rs
@@ -4,18 +4,24 @@ use self::compression::Compression;
 
 pub(crate) mod compression;
 
+/// Things supported by the node.
+///
+/// ### Layout
+///
+/// ```text
+///         24 bits              8 bits
+/// ┌─────────────────────┬──────────────────┐
+/// │     Compression     │     Reserved     │
+/// └─────────────────────┴──────────────────┘
+/// ```
+///
+/// 1. [`Compression`] - compression capabilities.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) struct Capabilities(u32);
 
-impl fmt::Display for Capabilities {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "(compression: {})", self.compression())
-    }
-}
-
 impl Capabilities {
     pub(crate) const fn new(compression: Compression) -> Self {
-        let compression = compression.bits() as u32;
+        let compression = compression.bits();
         let joined = compression << 8;
 
         Self(joined)
@@ -39,5 +45,63 @@ impl Capabilities {
 
     pub(crate) const fn bits(self) -> u32 {
         self.0
+    }
+}
+
+impl fmt::Display for Capabilities {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "(compression: {})", self.compression())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use self::compression::Algorithms;
+
+    #[test]
+    fn capabilities_format_is_compatible_with_020alpha17() {
+        let caps = Capabilities::new(Compression::new(Algorithms::LZ4, Algorithms::empty()));
+        let lz4_bit = caps.bits() & (1 << 8);
+
+        assert_eq!(lz4_bit, 1 << 8);
+    }
+
+    #[test]
+    fn compression_capabilities_encoded_right_way() {
+        #[track_caller]
+        fn case(create: (Algorithms, Algorithms), expect: (Algorithms, Algorithms)) {
+            let caps = Capabilities::new(Compression::new(create.0, create.1));
+            let compr = caps.compression();
+
+            assert_eq!(compr.supported(), expect.0);
+            assert_eq!(compr.preferred(), expect.1);
+
+            // Just in case we should decode same caps.
+
+            let bits = caps.bits();
+            let same_caps = Capabilities::from_bits_truncate(bits);
+
+            assert_eq!(caps, same_caps);
+        }
+
+        // Supported does not implies preferred.
+        case(
+            (Algorithms::LZ4, Algorithms::empty()),
+            (Algorithms::LZ4, Algorithms::empty()),
+        );
+
+        // Preferred implies supported.
+        case(
+            (Algorithms::empty(), Algorithms::LZ4),
+            (Algorithms::LZ4, Algorithms::LZ4),
+        );
+
+        // Nothing ever happens.
+        case(
+            (Algorithms::empty(), Algorithms::empty()),
+            (Algorithms::empty(), Algorithms::empty()),
+        );
     }
 }
